Add various rocket types with unique behaviors and visuals
- Implemented ArmoredRocket with health management and armor plating visuals.
- Created ArtilleryDesignator for targeting cities with a laser designation mechanic.
- Developed Drone class for agile movement and visual effects.
- Introduced FlareRocket to deploy flares at intervals.
- Added MirvRocket that splits into multiple child rockets upon reaching a certain height.
- Implemented ArtilleryShell for visual representation of artillery shells.
- Created StealthRocket with visibility toggling and distortion effects.
- Developed SwarmerRocket that splits into multiple drones upon reaching a certain height.
